ADD :: 데이터화하고 map을 써서 가독성을 높힘
